Timecode set() fps should be float.

+++ src/wx/timecode.h
@@ -61,7 +61,7 @@ public:
 
 	}
 
-	void set (T t, int fps)
+	void set (T t, float fps)
 	{
 		int h;
 		int m;
